% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamondDiabetesData.R
\name{DamondDiabetesData}
\alias{DamondDiabetesData}
\title{Obtain the Damond Diabetes data}
\usage{
DamondDiabetesData(...)
}
\value{
A \linkS4class{SegmentedCells} object.
}
\description{
Download and cache the Damond diabetes single-cell spatial (scSpatial) dataset from ExperimentHub,
returning a \linkS4class{SegmentedCells} object for further use.
}
\details{
This function provides the diabetes scSpatial data from Damond et al. (2019)
in the form of a \linkS4class{SegmentedCells} object.
}
\examples{
se <- DamondDiabetesData()

}
\references{
Damond, N., Engler, S., Zanotelli, V., Schapiro, D., Wasserfall, C., Kusmartseva, I., Nick, H., Thorel, F., Herrera, P., Atkinson, M. and Bodenmiller, B., 2019. A Map of Human Type 1 Diabetes Progression by Imaging Mass Cytometry.
\emph{Cell Metabolism}, 29(3), pp.755-768.e5.
}
\author{
Celso Milne
}
