% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HafnerBreastCancerData.R
\name{HafnerBreastCancerData}
\alias{HafnerBreastCancerData}
\title{Obtain the Hafner BreastCancer data}
\usage{
HafnerBreastCancerData(...)
}
\value{
A \linkS4class{SegmentedCells} object.
}
\description{
Download and cache the Hafner breast cancer single-cell spatial (scSpatial) dataset from ExperimentHub,
returning a \linkS4class{SegmentedCells} object for further use.
}
\details{
This function provides the breast cancer scSpatial data from Hafner et al. (2016)
in the form of a \linkS4class{SegmentedCells} object.
}
\examples{
se <- HafnerBreastCancerData()

}
\references{
Hafner, M., Niepel, M., Chung, M. and Sorger, P., 2016. Growth rate inhibition metrics correct for confounders in measuring sensitivity to cancer drugs.
\emph{Nature Methods},  13(6), pp.521-527. Available at: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4887336/}.
}
\author{
Celso Milne
}
