% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinMelanomaData.R
\name{LinMelanomaData}
\alias{LinMelanomaData}
\title{Obtain the Lin Melanoma data}
\usage{
LinMelanomaData()
}
\value{
A \linkS4class{SegmentedCells} object.
}
\description{
Download and cache the Lin melanoma single-cell spatial (scSpatial) dataset from ExperimentHub,
returning a \linkS4class{SegmentedCells} object for further use.
}
\details{
This function provides the melanoma scSpatial data from Lin et al. (2015)
in the form of a \linkS4class{SegmentedCells} object.
}
\examples{
se <- LinMelanomaData()

}
\references{
Lin, J., Fallahi-Sichani, M. and Sorger, P., 2015. Highly multiplexed imaging of single cells using a high-throughput cyclic immunofluorescence method.
\emph{Nature Communications}, 6(1). Available at: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4587398/}.
}
\author{
Celso Milne
}
